ai-agent/todo-agent-mcp/todo-mcp/pyproject.toml
[project]
name = "mcp-server-example"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp-server>=0.1.4",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "sqlite-utils>=3.38",
]
